Recursive dependencies
Now every node may be a dependency and graphs may have height more than 2 (dependencies may be added to another dependencies - that is what recursive dependencies mean).
Fix a bug in script of not checking if the pkg.out is present.

+++ Container/script.sh
@@ -15,7 +15,9 @@ if [ "$char" == "d" ]; then
 	if [ -e $dirname ]; then
 		cd "$dirname"
 
-		rm $curdir/pkg.out
+		if [ -e "$curdir/pkg.out" ]; then
+			rm $curdir/pkg.out
+		fi
 		for filename in ./*.spec
 		do
 			echo "Processing $filename file..."
